fix compiler warnings

+++ module/egress/src/EgressAsyncRunner.cpp
@@ -98,7 +98,7 @@ bool EgressAsyncRunner::Run(int argc, const char* const argv[], volatile bool & 
         std::cout << "EgressAsyncRunner: exiting cleanly..\n";
         egress.Stop();
         m_bundleCount = egress.m_telemetry.egressBundleCount;
-        m_bundleData = egress.m_telemetry.egressBundleData;
+        m_bundleData = static_cast<uint64_t>(egress.m_telemetry.egressBundleData);
         m_messageCount = egress.m_telemetry.egressMessageCount;
     }
     std::cout << "EgressAsyncRunner: exited cleanly\n";
